I'm tired, and beginning to lose my drive. I asked mr gpt to clean up the entire visuals so it's more readable and SOLID because I seriously cannot be bothered to do it myself. It being more layered like this is important for the aimbot that I'll do next.
