+++ Payday3-Internal_V2/Features/Visuals/VisualsHelpers.hpp
@@ -1,0 +1,17 @@
+#pragma once
+#include "pch.h"
+#include "VisualsTypes.hpp"
+
+namespace VisualsHelpers
+{
+	std::optional<ImVec4> CalculateScreenBoxFromTopBottom(SDK::APlayerController* pPlayerController, SDK::AActor* pActor,
+		SDK::FVector vecTop, SDK::FVector vecBottom);
+
+	std::optional<ImVec4> CalculateScreenBoxForCharacter(SDK::USkeletalMeshComponent* pMeshComponent,
+		SDK::APlayerController* pPlayerController, SDK::AActor* pActor);
+
+	void BuildBoneCache(SDK::USkeletalMeshComponent* mesh, VisualsTypes::BoneCache& cache);
+	void DrawBone(ImDrawList* pDrawList, SDK::APlayerController* pPlayerController, SDK::USkeletalMeshComponent* mesh,
+		int parent, int child, ImU32 color);
+
+}
